refactor(expenses): move custos operacionais para estrutura feature-first
