+++ front-end/src/features/expenses/constants/expense-options.ts
@@ -1,0 +1,11 @@
+import { CustomSelectOption } from '../../../shared/ui/Select';
+
+export const expenseCategoryOptions: CustomSelectOption[] = [
+  { value: 'Combustivel', label: 'Combustivel' },
+  { value: 'Manutencao', label: 'Manutencao' },
+  { value: 'Lavagem', label: 'Lavagem' },
+  { value: 'Pneus', label: 'Pneus' },
+  { value: 'Documentacao', label: 'Documentacao' },
+  { value: 'Seguro', label: 'Seguro' },
+  { value: 'Outros', label: 'Outros' },
+];
